requirements.txt
aiohttp
async-timeout
chardet
multidict
python-engineio
python-socketio
six
yarl
eventlet
asyncio
mavsdk
numpy
aioredis
redis
